+++ pkg/util/k8sutil/k8sutil.go
@@ -17,7 +17,6 @@ package k8sutil
 import (
 	"encoding/json"
 	"fmt"
-	"net/url"
 	"strings"
 	"time"
 
@@ -27,15 +26,18 @@ import (
 	"github.com/coreos/etcd-operator/pkg/util/etcdutil"
 	"github.com/coreos/etcd-operator/pkg/util/retryutil"
 
-	"k8s.io/client-go/1.5/kubernetes"
-	"k8s.io/client-go/1.5/pkg/api"
-	apierrors "k8s.io/client-go/1.5/pkg/api/errors"
-	"k8s.io/client-go/1.5/pkg/api/meta"
-	"k8s.io/client-go/1.5/pkg/api/meta/metatypes"
-	"k8s.io/client-go/1.5/pkg/api/v1"
-	"k8s.io/client-go/1.5/pkg/labels"
-	"k8s.io/client-go/1.5/pkg/util/intstr"
-	"k8s.io/client-go/1.5/rest"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/pkg/api"
+	apierrors "k8s.io/client-go/pkg/api/errors"
+	"k8s.io/client-go/pkg/api/meta"
+	"k8s.io/client-go/pkg/api/meta/metatypes"
+	"k8s.io/client-go/pkg/api/unversioned"
+	"k8s.io/client-go/pkg/api/v1"
+	"k8s.io/client-go/pkg/labels"
+	"k8s.io/client-go/pkg/runtime"
+	"k8s.io/client-go/pkg/runtime/serializer"
+	"k8s.io/client-go/pkg/util/intstr"
+	"k8s.io/client-go/rest"
 )
 
 const (
@@ -123,7 +125,7 @@ func BackupServiceName(clusterName string) string {
 }
 
 func CreateMemberService(kubecli kubernetes.Interface, ns string, svc *v1.Service) (*v1.Service, error) {
-	retSvc, err := kubecli.Core().Services(ns).Create(svc)
+	retSvc, err := kubecli.CoreV1().Services(ns).Create(svc)
 	if err != nil {
 		return nil, err
 	}
@@ -133,7 +135,7 @@ func CreateMemberService(kubecli kubernetes.Interface, ns string, svc *v1.Servic
 func CreateEtcdService(kubecli kubernetes.Interface, clusterName, ns string, owner metatypes.OwnerReference) (*v1.Service, error) {
 	svc := newEtcdServiceManifest(clusterName)
 	addOwnerRefToObject(svc.GetObjectMeta(), owner)
-	retSvc, err := kubecli.Core().Services(ns).Create(svc)
+	retSvc, err := kubecli.CoreV1().Services(ns).Create(svc)
 	if err != nil {
 		return nil, err
 	}
@@ -143,7 +145,7 @@ func CreateEtcdService(kubecli kubernetes.Interface, clusterName, ns string, own
 // CreateAndWaitPod is a workaround for self hosted and util for testing.
 // We should eventually get rid of this in critical code path and move it to test util.
 func CreateAndWaitPod(kubecli kubernetes.Interface, ns string, pod *v1.Pod, timeout time.Duration) (*v1.Pod, error) {
-	_, err := kubecli.Core().Pods(ns).Create(pod)
+	_, err := kubecli.CoreV1().Pods(ns).Create(pod)
 	if err != nil {
 		return nil, err
 	}
@@ -151,7 +153,7 @@ func CreateAndWaitPod(kubecli kubernetes.Interface, ns string, pod *v1.Pod, time
 	interval := 3 * time.Second
 	var retPod *v1.Pod
 	retryutil.Retry(interval, int(timeout/(interval)), func() (bool, error) {
-		retPod, err = kubecli.Core().Pods(ns).Get(pod.Name)
+		retPod, err = kubecli.CoreV1().Pods(ns).Get(pod.Name)
 		if err != nil {
 			return false, err
 		}
@@ -287,45 +289,33 @@ func NewEtcdPod(m *etcdutil.Member, initialCluster []string, clusterName, state,
 	return pod
 }
 
-func MustGetInClusterMasterHost() string {
+func MustNewKubeClient() kubernetes.Interface {
 	cfg, err := rest.InClusterConfig()
 	if err != nil {
 		panic(err)
 	}
-	return cfg.Host
+	return kubernetes.NewForConfigOrDie(cfg)
 }
 
-// tlsConfig isn't modified inside this function.
-// The reason it's a pointer is that it's not necessary to have tlsconfig to create a client.
-func MustCreateClient(host string, tlsInsecure bool, tlsConfig *rest.TLSClientConfig) kubernetes.Interface {
-	var cfg *rest.Config
-	if len(host) == 0 {
-		var err error
-		cfg, err = rest.InClusterConfig()
-		if err != nil {
-			panic(err)
-		}
-	} else {
-		cfg = &rest.Config{
-			Host:  host,
-			QPS:   100,
-			Burst: 100,
-		}
-		hostUrl, err := url.Parse(host)
-		if err != nil {
-			panic(fmt.Sprintf("failed to parse host url %s : %v", host, err))
-		}
-		if hostUrl.Scheme == "https" {
-			cfg.TLSClientConfig = *tlsConfig
-			cfg.Insecure = tlsInsecure
-		}
+func NewTPRClient() (*rest.RESTClient, error) {
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		return nil, err
 	}
 
-	c, err := kubernetes.NewForConfig(cfg)
-	if err != nil {
-		panic(err)
+	config.GroupVersion = &unversioned.GroupVersion{
+		Group:   spec.TPRGroup,
+		Version: spec.TPRVersion,
 	}
-	return c
+	config.APIPath = "/apis"
+	config.ContentType = runtime.ContentTypeJSON
+	config.NegotiatedSerializer = serializer.DirectCodecFactory{CodecFactory: api.Codecs}
+
+	restcli, err := rest.RESTClientFor(config)
+	if err != nil {
+		return nil, err
+	}
+	return restcli, nil
 }
 
 func IsKubernetesResourceAlreadyExistError(err error) bool {
@@ -337,9 +327,9 @@ func IsKubernetesResourceNotFoundError(err error) bool {
 }
 
 // We are using internal api types for cluster related.
-func ClusterListOpt(clusterName string) api.ListOptions {
-	return api.ListOptions{
-		LabelSelector: labels.SelectorFromSet(newLablesForCluster(clusterName)),
+func ClusterListOpt(clusterName string) v1.ListOptions {
+	return v1.ListOptions{
+		LabelSelector: labels.SelectorFromSet(newLablesForCluster(clusterName)).String(),
 	}
 }